feat: Add tsconfig and refactor code into different ts files
